move array, inProver
